Move blending stuff to BlendingModeParameters.cs.
