+++ osu.Framework/osu.Framework.csproj
@@ -115,6 +115,7 @@
     <Compile Include="Graphics\Animations\FrameData.cs" />
     <Compile Include="Graphics\Animations\IAnimation.cs" />
     <Compile Include="Graphics\Animations\TextureAnimation.cs" />
+    <Compile Include="Graphics\BlendingModeParameters.cs" />
     <Compile Include="Graphics\Containers\AsyncLoadWrapper.cs" />
     <Compile Include="Graphics\Containers\CompositeDrawable.cs" />
     <Compile Include="Graphics\Containers\CustomizableTextContainer.cs" />
